Исправление директорий
