+++ backend/src/main/java/mediasoft/ru/backend/models/dto/request/product/UpdateProductRequestDTO.java
@@ -1,4 +1,4 @@
-package mediasoft.ru.backend.dto.response.product;
+package mediasoft.ru.backend.models.dto.request.product;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -8,8 +8,6 @@ import lombok.Setter;
 import mediasoft.ru.backend.enums.ProductCategory;
 
 import java.math.BigDecimal;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.UUID;
 
 @Getter
@@ -17,7 +15,7 @@ import java.util.UUID;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ProductInfoResponseDTO {
+public class UpdateProductRequestDTO {
     private UUID id;
     private String article;
     private String name;
@@ -25,6 +23,4 @@ public class ProductInfoResponseDTO {
     private ProductCategory category;
     private BigDecimal price;
     private BigDecimal count;
-    private LocalDate creationDate;
-    private LocalDateTime lastModifiedDate;
 }
