+++ backend/src/main/java/mediasoft/ru/backend/models/dto/response/product/ProductInfoResponseDTO.java
@@ -1,4 +1,4 @@
-package mediasoft.ru.backend.dto.response.product;
+package mediasoft.ru.backend.models.dto.response.product;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -8,6 +8,7 @@ import lombok.Setter;
 import mediasoft.ru.backend.enums.ProductCategory;
 
 import java.math.BigDecimal;
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -16,7 +17,7 @@ import java.util.UUID;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class UpdateProductResponseDTO {
+public class ProductInfoResponseDTO {
     private UUID id;
     private String article;
     private String name;
@@ -24,6 +25,6 @@ public class UpdateProductResponseDTO {
     private ProductCategory category;
     private BigDecimal price;
     private BigDecimal count;
+    private LocalDate creationDate;
     private LocalDateTime lastModifiedDate;
 }
-
