+++ backend/src/main/java/mediasoft/ru/backend/models/dto/response/product/UpdateProductResponseDTO.java
@@ -1,4 +1,4 @@
-package mediasoft.ru.backend.dto;
+package mediasoft.ru.backend.models.dto.response.product;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -8,7 +8,6 @@ import lombok.Setter;
 import mediasoft.ru.backend.enums.ProductCategory;
 
 import java.math.BigDecimal;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -17,7 +16,7 @@ import java.util.UUID;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ProductDTO {
+public class UpdateProductResponseDTO {
     private UUID id;
     private String article;
     private String name;
@@ -25,6 +24,6 @@ public class ProductDTO {
     private ProductCategory category;
     private BigDecimal price;
     private BigDecimal count;
-    private LocalDate creationDate;
     private LocalDateTime lastModifiedDate;
 }
+
